------------------------------------------------------------------------ r6627 | malc | 2009-02-18 21:44:04 +0100 (Mi, 18 Feb 2009) | 1 line
Avoid running audio ctl's when vm is not running
------------------------------------------------------------------------
r6628 | aurel32 | 2009-02-18 22:37:17 +0100 (Mi, 18 Feb 2009) | 16 lines

Fix cpu_physical_memory_rw() for 64-bit I/O accesses

KVM uses cpu_physical_memory_rw() to access the I/O devices. When a
read or write with a length of 8-byte is requested, it is split into 2
4-byte accesses.

This has been broken in revision 5849. After this revision, only the
first 4 bytes are actually read/write to the device, as the target
address is changed, so on the next iteration of the loop the next 4
bytes are actually read/written elsewhere (in the RAM for the graphic
card).

This patch fixes screen corruption (and most probably data corruption)
with FreeBSD/amd64. Bug #2556746 in KVM bugzilla.

Signed-off-by: Aurelien Jarno <aurelien@aurel32.net>
------------------------------------------------------------------------