------------------------------------------------------------------------ r6996 | aliguori | 2009-04-05 21:29:26 +0200 (So, 05. Apr 2009) | 7 lines
Disable qemu-io on Win32

It breaks the build.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>


------------------------------------------------------------------------
r6997 | aliguori | 2009-04-05 21:30:33 +0200 (So, 05. Apr 2009) | 12 lines

propagate error on failed completion (Glauber Costa)

migrate_fd_put_ready() calls qemu_savevm_state_complete(),
but the later can fail.

If it happens, re-start the vm and propagate the error up

Based on a patch by Yaniv Kamay

Signed-off-by: Glauber Costa <glommer@redhat.com>
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>

------------------------------------------------------------------------
r6998 | aliguori | 2009-04-05 21:30:51 +0200 (So, 05. Apr 2009) | 13 lines

create qemu_file_set_error (Glauber Costa)

This is mainly for consistency, since we don't want
anything outside of savevm setting it explicitly. There
are current no users of that in qemu tree, but there
are potential candidates on kvm-userspace. And avi
is a nice guy, let's be nice with him.

Based on a patch by Yaniv Kamay

Signed-off-by: Glauber Costa <glommer@redhat.com>
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>

------------------------------------------------------------------------
r6999 | aliguori | 2009-04-05 21:30:55 +0200 (So, 05. Apr 2009) | 12 lines

stop dirty tracking just at the end of migration (Glauber Costa)

If there is still work to do, it is not safe to assume we
can end the dirty tracking. Specifically, kvm can update the dirty
tracking log inside ram_save_block(), leaving pages still out of sync
if we go with the current code.

Based on a patch by Yaniv Kamay

Signed-off-by: Glauber Costa <glommer@redhat.com>
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>

------------------------------------------------------------------------